requirements.txt
fastapi
uvicorn[standard]
jinja2
sqlalchemy
python-dotenv
passlib[bcrypt]
python-multipart
openai
anthropic
itsdangerous
resend
